#1948 :recycle: move uuid logic to utils
