+++ wls-gui-wahllokalsystem/src/composables/common/textFormatter.ts
@@ -20,7 +20,19 @@ export function useTextFormatter() {
     }
   }
 
+  function createUuidv4() {
+    return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(
+      /[xy]/g,
+      function (c) {
+        const r = (Math.random() * 16) | 0;
+        const v = c == "x" ? r : (r & 0x3) | 0x8;
+        return v.toString(16);
+      }
+    );
+  }
+
   return {
     getStimmzettelTermForWahl,
+    createUuidv4,
   };
 }
